backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.0
pydantic-settings==2.1.0
openai==2.33.0
tiktoken==0.12.0
langchain==0.0.340
langchain-community==0.0.10
chromadb==0.4.22
pypdf2==3.0.1
python-docx==1.2.0
beautifulsoup4==4.14.3
markdown==3.10.2
python-magic==0.4.27
pgvector==0.2.4
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
httpx==0.25.0
psutil==5.9.0
redis==5.0.0
aioredis==2.0.1
celery==5.3.4
flower==2.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6